+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Support sorting NULLs first or last
 * Added JSON types
 * Added `runInsertMany`